move Render implementation into child classes
The bar plot base now contains data management only.

This commit leaves child classes with sub-ideal duplicated code. Plot-type-specific minimal-complexity render methods should be created for these child classes in the future.
