+++ src/ScottPlot/Plottable/BarPlotBase.cs
@@ -1,14 +1,9 @@
-﻿using ScottPlot.Drawing;
-using System;
-using System.Collections.Generic;
+﻿using System;
 using System.Drawing;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace ScottPlot.Plottable
 {
-    public abstract class BarPlotBase : IPlottable
+    public abstract class BarPlotBase
     {
         public bool IsVisible { get; set; } = true;
         public int XAxisIndex { get; set; } = 0;
@@ -59,27 +54,29 @@ namespace ScottPlot.Plottable
         public double ValueBase { get; set; }
 
         /// <summary>
-        /// Width of bars (axis units)
+        /// Width of bars defined in axis units.
+        /// If bars are evenly spaced, consider setting this to a fraction of the distance between the first two Positions.
         /// </summary>
         public double BarWidth = .8;
 
         /// <summary>
-        /// Width of the errorbars (axis units)
+        /// Width of the errorbar caps defined in axis units.
         /// </summary>
         public double ErrorCapSize = .4;
 
         /// <summary>
-        /// Thickness of the errorbars (pixel units)
+        /// Thickness of the errorbar lines (pixel units)
         /// </summary>
         public float ErrorLineWidth = 1;
 
         /// <summary>
-        /// Outline each bar with this color. Set to transparent to disable outlines.
+        /// Outline each bar with this color. 
+        /// Set this to transparent to disable outlines.
         /// </summary>
         public Color BorderColor = Color.Black;
 
         /// <summary>
-        /// Color of errorbars.
+        /// Color of errorbar lines.
         /// </summary>
         public Color ErrorColor = Color.Black;
 
@@ -120,126 +117,12 @@ namespace ScottPlot.Plottable
                 new AxisLimits(valueMin, valueMax, positionMin, positionMax);
         }
 
-        public abstract LegendItem[] GetLegendItems();
-
-        public void Render(PlotDimensions dims, Bitmap bmp, bool lowQuality = false)
-        {
-            using Graphics gfx = GDI.Graphics(bmp, dims, lowQuality);
-            for (int barIndex = 0; barIndex < Values.Length; barIndex++)
-            {
-                if (Orientation == Orientation.Vertical)
-                    RenderBarVertical(dims, gfx, Positions[barIndex] + XOffset, Values[barIndex], ValueErrors[barIndex], ValueOffsets[barIndex]);
-                else
-                    RenderBarHorizontal(dims, gfx, Positions[barIndex] + XOffset, Values[barIndex], ValueErrors[barIndex], ValueOffsets[barIndex]);
-            }
-        }
-
-        protected abstract void RenderBarFromRect(RectangleF rect, bool negative, Graphics gfx);
-
-        private void RenderBarVertical(PlotDimensions dims, Graphics gfx, double position, double value, double valueError, double yOffset)
-        {
-            // bar body
-            float centerPx = dims.GetPixelX(position);
-            double edge1 = position - BarWidth / 2;
-            double value1 = Math.Min(ValueBase, value) + yOffset;
-            double value2 = Math.Max(ValueBase, value) + yOffset;
-            double valueSpan = value2 - value1;
-
-            var rect = new RectangleF(
-                x: dims.GetPixelX(edge1),
-                y: dims.GetPixelY(value2),
-                width: (float)(BarWidth * dims.PxPerUnitX),
-                height: (float)(valueSpan * dims.PxPerUnitY));
-
-            // errorbar
-            double error1 = value > 0 ? value2 - Math.Abs(valueError) : value1 - Math.Abs(valueError);
-            double error2 = value > 0 ? value2 + Math.Abs(valueError) : value1 + Math.Abs(valueError);
-            float capPx1 = dims.GetPixelX(position - ErrorCapSize * BarWidth / 2);
-            float capPx2 = dims.GetPixelX(position + ErrorCapSize * BarWidth / 2);
-            float errorPx2 = dims.GetPixelY(error2);
-            float errorPx1 = dims.GetPixelY(error1);
-
-            RenderBarFromRect(rect, value < 0, gfx);
-
-            if (ErrorLineWidth > 0 && valueError > 0)
-            {
-                using var errorPen = new Pen(ErrorColor, ErrorLineWidth);
-                gfx.DrawLine(errorPen, centerPx, errorPx1, centerPx, errorPx2);
-                gfx.DrawLine(errorPen, capPx1, errorPx1, capPx2, errorPx1);
-                gfx.DrawLine(errorPen, capPx1, errorPx2, capPx2, errorPx2);
-            }
-
-            if (ShowValuesAboveBars)
-                using (var valueTextFont = GDI.Font(Font))
-                using (var valueTextBrush = GDI.Brush(Font.Color))
-                using (var sf = new StringFormat() { LineAlignment = StringAlignment.Far, Alignment = StringAlignment.Center })
-                    gfx.DrawString(value.ToString(), valueTextFont, valueTextBrush, centerPx, rect.Y, sf);
-        }
-
-        private void RenderBarHorizontal(PlotDimensions dims, Graphics gfx, double position, double value, double valueError, double yOffset)
-        {
-            // bar body
-            float centerPx = dims.GetPixelY(position);
-            double edge2 = position + BarWidth / 2;
-            double value1 = Math.Min(ValueBase, value) + yOffset;
-            double value2 = Math.Max(ValueBase, value) + yOffset;
-            double valueSpan = value2 - value1;
-            var rect = new RectangleF(
-                x: dims.GetPixelX(value1),
-                y: dims.GetPixelY(edge2),
-                height: (float)(BarWidth * dims.PxPerUnitY),
-                width: (float)(valueSpan * dims.PxPerUnitX));
-
-            RenderBarFromRect(rect, value < 0, gfx);
-
-            // errorbar
-            double error1 = value > 0 ? value2 - Math.Abs(valueError) : value1 - Math.Abs(valueError);
-            double error2 = value > 0 ? value2 + Math.Abs(valueError) : value1 + Math.Abs(valueError);
-            float capPx1 = dims.GetPixelY(position - ErrorCapSize * BarWidth / 2);
-            float capPx2 = dims.GetPixelY(position + ErrorCapSize * BarWidth / 2);
-            float errorPx2 = dims.GetPixelX(error2);
-            float errorPx1 = dims.GetPixelX(error1);
-
-            if (ErrorLineWidth > 0 && valueError > 0)
-            {
-                using var errorPen = new Pen(ErrorColor, ErrorLineWidth);
-                gfx.DrawLine(errorPen, errorPx1, centerPx, errorPx2, centerPx);
-                gfx.DrawLine(errorPen, errorPx1, capPx2, errorPx1, capPx1);
-                gfx.DrawLine(errorPen, errorPx2, capPx2, errorPx2, capPx1);
-            }
-
-            if (ShowValuesAboveBars)
-                using (var valueTextFont = GDI.Font(Font))
-                using (var valueTextBrush = GDI.Brush(Font.Color))
-                using (var sf = new StringFormat() { LineAlignment = StringAlignment.Center, Alignment = StringAlignment.Near })
-                    gfx.DrawString(value.ToString(), valueTextFont, valueTextBrush, rect.X + rect.Width, centerPx, sf);
-        }
-
-        public void ValidateData(bool deep = false)
-        {
-            Validate.AssertHasElements("xs", Positions);
-            Validate.AssertHasElements("ys", Values);
-            Validate.AssertHasElements("yErr", ValueErrors);
-            Validate.AssertHasElements("yOffsets", ValueOffsets);
-            Validate.AssertEqualLength("xs, ys, yErr, and yOffsets", Positions, Values, ValueErrors, ValueOffsets);
-
-            if (deep)
-            {
-                Validate.AssertAllReal("xs", Positions);
-                Validate.AssertAllReal("ys", Values);
-                Validate.AssertAllReal("yErr", ValueErrors);
-                Validate.AssertAllReal("yOffsets", ValueOffsets);
-            }
-
-        }
-
         [Obsolete("Reference the 'Orientation' field instead of this field")]
         public bool VerticalOrientation
         {
             get => Orientation == Orientation.Vertical;
             set => Orientation = value ? Orientation.Vertical : Orientation.Horizontal;
         }
-
 
         [Obsolete("Reference the 'Orientation' field instead of this field")]
         public bool HorizontalOrientation
