+++ src/ScottPlot/Plottable/ClevelandDotPlot.cs
@@ -13,7 +13,7 @@ namespace ScottPlot.Plottable
     /// Positions are defined by Xs.
     /// Heights are defined by Ys1 and Ys2 (internally done with Ys and YOffsets).
     /// </summary>
-    public class ClevelandDotPlot : BarPlotBase
+    public class ClevelandDotPlot : BarPlotBase, IPlottable
     {
         /// <summary>
         /// Color for the line
@@ -153,7 +153,7 @@ namespace ScottPlot.Plottable
                 new AxisLimits(valueMin, valueMax, positionMin, positionMax);
         }
 
-        public override LegendItem[] GetLegendItems()
+        public LegendItem[] GetLegendItems()
         {
             var firstDot = new LegendItem()
             {
@@ -175,7 +175,24 @@ namespace ScottPlot.Plottable
             return new LegendItem[] { firstDot, secondDot };
         }
 
-        protected override void RenderBarFromRect(RectangleF rect, bool negative, Graphics gfx)
+        #region Render Implementation
+
+        // NOTE: These render methods contains a lot of code and complexity not required to render this plot type.
+        // TODO: Delete as much of this code as possible and simplify the render methods.
+
+        public void Render(PlotDimensions dims, Bitmap bmp, bool lowQuality = false)
+        {
+            using Graphics gfx = GDI.Graphics(bmp, dims, lowQuality);
+            for (int barIndex = 0; barIndex < Values.Length; barIndex++)
+            {
+                if (Orientation == Orientation.Vertical)
+                    RenderBarVertical(dims, gfx, Positions[barIndex] + XOffset, Values[barIndex], ValueErrors[barIndex], ValueOffsets[barIndex]);
+                else
+                    RenderBarHorizontal(dims, gfx, Positions[barIndex] + XOffset, Values[barIndex], ValueErrors[barIndex], ValueOffsets[barIndex]);
+            }
+        }
+
+        private void RenderBarFromRect(RectangleF rect, bool negative, Graphics gfx)
         {
             float centerPx = Orientation == Orientation.Horizontal
                 ? rect.Y + rect.Height / 2
@@ -199,6 +216,92 @@ namespace ScottPlot.Plottable
             gfx.DrawLine(stemPen, points[0], points[1]);
             MarkerTools.DrawMarker(gfx, points[1], MarkerShape2, DotRadius, Color2);
             MarkerTools.DrawMarker(gfx, points[0], MarkerShape1, DotRadius, Color1); // First point should be drawn overtop the second.
+        }
+
+        private void RenderBarVertical(PlotDimensions dims, Graphics gfx, double position, double value, double valueError, double yOffset)
+        {
+            // bar body
+            float centerPx = dims.GetPixelX(position);
+            double edge1 = position - BarWidth / 2;
+            double value1 = Math.Min(ValueBase, value) + yOffset;
+            double value2 = Math.Max(ValueBase, value) + yOffset;
+            double valueSpan = value2 - value1;
+
+            var rect = new RectangleF(
+                x: dims.GetPixelX(edge1),
+                y: dims.GetPixelY(value2),
+                width: (float)(BarWidth * dims.PxPerUnitX),
+                height: (float)(valueSpan * dims.PxPerUnitY));
+
+            // errorbar
+            double error1 = value > 0 ? value2 - Math.Abs(valueError) : value1 - Math.Abs(valueError);
+            double error2 = value > 0 ? value2 + Math.Abs(valueError) : value1 + Math.Abs(valueError);
+            float capPx1 = dims.GetPixelX(position - ErrorCapSize * BarWidth / 2);
+            float capPx2 = dims.GetPixelX(position + ErrorCapSize * BarWidth / 2);
+            float errorPx2 = dims.GetPixelY(error2);
+            float errorPx1 = dims.GetPixelY(error1);
+
+            RenderBarFromRect(rect, value < 0, gfx);
+
+            if (ErrorLineWidth > 0 && valueError > 0)
+            {
+                using var errorPen = new Pen(ErrorColor, ErrorLineWidth);
+                gfx.DrawLine(errorPen, centerPx, errorPx1, centerPx, errorPx2);
+                gfx.DrawLine(errorPen, capPx1, errorPx1, capPx2, errorPx1);
+                gfx.DrawLine(errorPen, capPx1, errorPx2, capPx2, errorPx2);
+            }
+
+            if (ShowValuesAboveBars)
+                using (var valueTextFont = GDI.Font(Font))
+                using (var valueTextBrush = GDI.Brush(Font.Color))
+                using (var sf = new StringFormat() { LineAlignment = StringAlignment.Far, Alignment = StringAlignment.Center })
+                    gfx.DrawString(value.ToString(), valueTextFont, valueTextBrush, centerPx, rect.Y, sf);
+        }
+
+        private void RenderBarHorizontal(PlotDimensions dims, Graphics gfx, double position, double value, double valueError, double yOffset)
+        {
+            // bar body
+            float centerPx = dims.GetPixelY(position);
+            double edge2 = position + BarWidth / 2;
+            double value1 = Math.Min(ValueBase, value) + yOffset;
+            double value2 = Math.Max(ValueBase, value) + yOffset;
+            double valueSpan = value2 - value1;
+            var rect = new RectangleF(
+                x: dims.GetPixelX(value1),
+                y: dims.GetPixelY(edge2),
+                height: (float)(BarWidth * dims.PxPerUnitY),
+                width: (float)(valueSpan * dims.PxPerUnitX));
+
+            RenderBarFromRect(rect, value < 0, gfx);
+
+            // errorbar
+            double error1 = value > 0 ? value2 - Math.Abs(valueError) : value1 - Math.Abs(valueError);
+            double error2 = value > 0 ? value2 + Math.Abs(valueError) : value1 + Math.Abs(valueError);
+            float capPx1 = dims.GetPixelY(position - ErrorCapSize * BarWidth / 2);
+            float capPx2 = dims.GetPixelY(position + ErrorCapSize * BarWidth / 2);
+            float errorPx2 = dims.GetPixelX(error2);
+            float errorPx1 = dims.GetPixelX(error1);
+
+            if (ErrorLineWidth > 0 && valueError > 0)
+            {
+                using var errorPen = new Pen(ErrorColor, ErrorLineWidth);
+                gfx.DrawLine(errorPen, errorPx1, centerPx, errorPx2, centerPx);
+                gfx.DrawLine(errorPen, errorPx1, capPx2, errorPx1, capPx1);
+                gfx.DrawLine(errorPen, errorPx2, capPx2, errorPx2, capPx1);
+            }
+
+            if (ShowValuesAboveBars)
+                using (var valueTextFont = GDI.Font(Font))
+                using (var valueTextBrush = GDI.Brush(Font.Color))
+                using (var sf = new StringFormat() { LineAlignment = StringAlignment.Center, Alignment = StringAlignment.Near })
+                    gfx.DrawString(value.ToString(), valueTextFont, valueTextBrush, rect.X + rect.Width, centerPx, sf);
+        }
+
+        #endregion
+
+        public void ValidateData(bool deep = false)
+        {
+            // TODO: refactor entire data validation system for all plot types (triaged March 2021)
         }
     }
 }
